fw/services/bluetooth: do not choke on gap_bondings of the wrong version
snowy uses v1 of bluetooth_persistent_storage, which NimBLE does not
support.  As it turns out, we are generally not very robust against
gap_bondings that we don't know how to deal with, and we can crash the
system if they happen!  Let's rather ignore them, in case a future PebbleOS
version extends the gap_bonding structure in a way that we don't know how to
read and someone tries to downgrade.

Signed-off-by: Joshua Wise <joshua@joshuawise.com>